Improved random firing code which uses the number of frames as reference to determine firing interval. This works better since speed of firing can be changed to as low as once per frame. The higher the frame rate setting, the more fine grained the firing speed will be.

+++ OuterSpace/Models/AiModel.cs
@@ -15,7 +15,7 @@ namespace OuterSpace.Models
         public int HeadingRange { get; set; }
         public string RandomStart { get; set; }
         public int Weapon { get; set; }
-        public int FireGranularity { get; set; }
+        public double FireGranularity { get; set; }
         public int Speed { get; set; }
         public double ScanRange { get; set; }
         public bool HitBarShow { get; set; }
